Se añade el zoom inicial

+++ app/src/main/java/com/primos/visitamoraleja/contenidos/Evento.java
@@ -23,7 +23,7 @@ public class Evento implements Serializable {
 	private Bitmap icono;
 	private double longitud;
 	private double latitud;
-	private int zoomInicial;
+	private float zoomInicial;
 	private Date inicio;
 	private Date fin;
 	private boolean activo;
@@ -36,11 +36,11 @@ public class Evento implements Serializable {
 	}
 
 
-	public int getZoomInicial() {
+	public float getZoomInicial() {
 		return zoomInicial;
 	}
 
-	public void setZoomInicial(int zoomInicial) {
+	public void setZoomInicial(float zoomInicial) {
 		this.zoomInicial = zoomInicial;
 	}
